Added menu items with options to homepage. Very primitive/naive implementation, will update

+++ app/controllers/home_controller.rb
@@ -6,7 +6,7 @@ class HomeController < ApplicationController
   def top_menu_items
     render :json => {:success => true, :data => MenuItem.find_top_menu_items(params[:latitude].to_f,
                                                                              params[:longitude].to_f,
-                                                                             params[:range].to_i)[0..(params[:limit].to_i - 1)]}
+                                                                             params[:range].to_f)[0..(params[:limit].to_i - 1)]}
   end
 
   def search_by_user_selected_category
